+++ frontend/src/pages/Traces/TracesList.tsx
@@ -18,11 +18,11 @@ import { useVirtualizer } from '@tanstack/react-virtual'
 import { isEqual } from 'lodash'
 import React, { Key, useMemo, useRef } from 'react'
 
-import { CustomColumnActions } from '@/components/CustomColumnActions'
 import {
-	CustomColumnPopover,
-	TraceCustomColumn,
-} from '@/components/CustomColumnPopover'
+	ColumnHeader,
+	CustomColumnHeader,
+} from '@/components/CustomColumnHeader'
+import { CustomColumnPopover } from '@/components/CustomColumnPopover'
 import { AdditionalFeedResults } from '@/components/FeedResults/FeedResults'
 import { LinkButton } from '@/components/LinkButton'
 import LoadingBox from '@/components/LoadingBox'
@@ -44,13 +44,6 @@ type Props = {
 	resetMoreTraces?: () => void
 	fetchMoreWhenScrolled: (target: HTMLDivElement) => void
 	loadingAfter: boolean
-}
-
-type ColumnHeader = {
-	id: string
-	component: React.ReactNode
-	showActions?: boolean
-	noPadding?: boolean
 }
 
 const LOADING_AFTER_HEIGHT = 28
@@ -228,11 +221,12 @@ export const TracesList: React.FC<Props> = ({
 			<Table.Head>
 				<Table.Row gridColumns={columnData.gridColumns}>
 					{columnData.columnHeaders.map((header) => (
-						<TableHeader
+						<CustomColumnHeader
 							key={header.id}
 							header={header}
 							selectedColumns={selectedColumns}
-							setSelectedColumns={setSelectedColumns}
+							setSelectedColumns={setSelectedColumns!}
+							standardColumns={HIGHLIGHT_STANDARD_COLUMNS}
 						/>
 					))}
 				</Table.Row>
@@ -340,95 +334,3 @@ const TracesTableRow = React.memo<TracesTableRowProps>(
 		)
 	},
 )
-
-type TableHeaderProps = {
-	header: ColumnHeader
-	selectedColumns: TraceCustomColumn[]
-	setSelectedColumns: (columns: TraceCustomColumn[]) => void
-}
-
-const TableHeader: React.FC<TableHeaderProps> = ({
-	header,
-	selectedColumns,
-	setSelectedColumns,
-}) => {
-	const headerRef = useRef<HTMLDivElement>(null)
-
-	const columnIndex = useMemo(
-		() => selectedColumns.findIndex((c) => c.id === header.id),
-		[selectedColumns, header.id],
-	)
-
-	const handleDrag = (e: React.MouseEvent) => {
-		const leftElementCoord = headerRef.current?.getBoundingClientRect()
-		const rightElementCoord =
-			headerRef.current?.nextElementSibling?.getBoundingClientRect()
-
-		if (
-			leftElementCoord &&
-			rightElementCoord &&
-			e.pageX > leftElementCoord.left &&
-			e.pageX < rightElementCoord.right
-		) {
-			const leftElementWidth = leftElementCoord.width
-			const rightElementWidth = rightElementCoord.width
-			const leftElementNewWidth = e.pageX - leftElementCoord.left
-			const rightElementNewWidth =
-				rightElementWidth - (leftElementNewWidth - leftElementWidth)
-
-			const newSelectedColumns = [...selectedColumns]
-			newSelectedColumns[columnIndex] = {
-				...newSelectedColumns[columnIndex],
-				size: `${leftElementNewWidth}px`,
-			}
-			newSelectedColumns[columnIndex + 1] = {
-				...newSelectedColumns[columnIndex + 1],
-				size: `${rightElementNewWidth}px`,
-			}
-
-			setSelectedColumns(newSelectedColumns)
-		}
-	}
-
-	const resizeable =
-		header.showActions && columnIndex !== selectedColumns.length - 1
-
-	return (
-		<Table.Header
-			key={header.id}
-			noPadding={header.noPadding}
-			ref={headerRef}
-		>
-			<Box
-				display="flex"
-				alignItems="center"
-				justifyContent="space-between"
-			>
-				<Text lines="1">{header.component}</Text>
-				{header.showActions && (
-					<CustomColumnActions
-						columnId={header.id}
-						selectedColumns={selectedColumns}
-						setSelectedColumns={setSelectedColumns}
-						trackingId="TracesTableColumn"
-						standardColumns={HIGHLIGHT_STANDARD_COLUMNS}
-					/>
-				)}
-			</Box>
-			{resizeable && (
-				<Box
-					style={{
-						position: 'absolute',
-						top: 0,
-						bottom: 0,
-						right: 0,
-						width: 3,
-						cursor: 'col-resize',
-					}}
-					draggable
-					onDrag={handleDrag}
-				/>
-			)}
-		</Table.Header>
-	)
-}
